Set memory/msg/ready in %post to fix msg-server state detection
When msg-server launches, it sets memory/msg/ready to True, but if the
key does not already exist, this set will fail.

Also, memory status reports (from Settings), Messages app settings, and
the Messages app itself rely on this vconf key being set to True to
function properly.

Create the vconf key in %post for this package to fix the issue.

Change-Id: I4c12a41d84a8d2a284aa8f8e81655cba79a702be
Signed-off-by: Patrick McCarty <patrick.mccarty@linux.intel.com>