Set memory/msg/ready in %post to fix msg-server state detection
When msg-server launches, it sets memory/msg/ready to True, but if the
key does not already exist, this set will fail.

Also, memory status reports (from Settings), Messages app settings, and
the Messages app itself rely on this vconf key being set to True to
function properly.

Create the vconf key in %post for this package to fix the issue.

Signed-off-by: Patrick McCarty <patrick.mccarty@linux.intel.com>

Conflicts:

	packaging/msg-service.changes

Change-Id: Iead53b2085d8f7e23c2ab5d4209b5f4d511c176e